remove anyhow; simple unit tests
