+++ Cargo.toml
@@ -9,7 +9,6 @@ lto = true
 codegen-units = 1
 
 [dependencies]
-anyhow = "1.0.70"
 axum = "0.6.16"
 genius-rust = "0.1.1"
 http = "0.2.9"
